amend: refactoring
